pyproject.toml
[project]
name = "atividade_final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf>=1.7.2",
    "google-genai>=1.52.0",
    "google-generativeai>=0.8.5",
    "markdown>=3.7",
    "openai-whisper>=20250625",
    "pypdf>=6.3.0",
    "streamlit>=1.51.0",
    "weasyprint>=63.1",
]
